+++ src/print/print_u.c
@@ -6,13 +6,13 @@
 /*   By: xvoorvaa <xvoorvaa@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/06/07 11:27:44 by xvoorvaa      #+#    #+#                 */
-/*   Updated: 2021/07/08 16:59:50 by xander        ########   odam.nl         */
+/*   Updated: 2021/07/08 18:56:52 by xander        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../printf.h"
 
-size_t	print_u_normal(unsigned long long hex)
+int	print_u_normal(unsigned long long hex)
 {
 	int	i;
 	int	len;
@@ -26,12 +26,12 @@ size_t	print_u_normal(unsigned long long hex)
 		write(1, "0", 1);
 		i++;
 	}
-	if (t_flags.precision > -1  || numlen(hex) > 1)
+	if (t_flags.precision > -1 || numlen(hex) > 1)
 		len = ft_putlong_fd(hex, 1, 0) + len;
 	return (len + i);
 }
 
-size_t	print_u_minus(unsigned long long hex)
+int	print_u_minus(unsigned long long hex)
 {
 	int	i;
 	int	len;
@@ -49,7 +49,7 @@ size_t	print_u_minus(unsigned long long hex)
 	return (len + i);
 }
 
-size_t	print_u(va_list ap)
+int	print_u(va_list ap)
 {
 	size_t				len;
 	unsigned long long	hex;
